Applied my existing edits to the 2.6.29 kernel of the new bitbake/openembedded environment. The kernel still boots and also loads the helloMcBSP.ko module and outputs the correct signals on the Expansion port. It also receives valid data from the ADS1258 until the McBSPs FIFO buffer is full (128 bytes), after which there is a break in the data. So after testing that this also can be applied to the 2.6.31 kernel, I will add another method to mcbsp.c to solve the buffering issue.
git-svn-id: file:///Users/simon/Workspaces/WorkspaceSVN/ISIP_svn/svn/beagleboard/trunk@23 1a49bced-557f-0410-a62c-c1dc85da7da9

+++ kernelChanges/bitbake-default-linux-omap-2.6.29/arch/arm/plat-omap/include/mach/mcbsp.h
@@ -424,8 +424,8 @@ int omap_mcbsp_spi_master_recv_word_poll(unsigned int id, u32 * word);
 void omap_mcbsp_set_spi_mode(unsigned int id, const struct omap_mcbsp_spi_cfg * spi_cfg);
 
 /* Polled read/write functions */
-int omap_mcbsp_pollread(unsigned int id, u16 * buf);
-int omap_mcbsp_pollwrite(unsigned int id, u16 buf);
+int omap_mcbsp_pollread(unsigned int id, u32 * buf);
+int omap_mcbsp_pollwrite(unsigned int id, u32 buf);
 int omap_mcbsp_set_io_type(unsigned int id, omap_mcbsp_io_type_t io_type);
 
 #endif
